get blocks from fixtures
