+++ ol/fixtures/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2018"
 
 [dependencies]
 ol-types = {path = "../types"}
+serde_json = "1.0.64"
