+++ ol/fixtures/Makefile
@@ -17,18 +17,16 @@ check:
 		PERSONA=$$i MNEM=$$(cat ${SOURCE}/ol/fixtures/mnemonic/$$i.mnem) make check; \
 	done
 
-# MNEM = $(shell cat ${SOURCE}/ol/fixtures/mnemonic/${PERSONA}.mnem)
-
 echo: 
 	@echo ${PERSONA}:
 	@echo ${MNEM}
 
 fixtures-test: 
-	cd ../onboard && cargo r -p onboard -- val --genesis-ceremony --autopay-file ${SOURCE}/ol/fixtures/autopay/${PERSONA}.autopay_batch.json
+	cd ../onboard && NODE_ENV=test TEST=y cargo r -p onboard -- val --genesis-ceremony --autopay-file ${SOURCE}/ol/fixtures/autopay/${PERSONA}.autopay_batch.json
 	cp ${DATA_PATH}/blocks/block_0.json ${SOURCE}/ol/fixtures/blocks/test/${PERSONA}/
 	cp ${DATA_PATH}/account.json ${SOURCE}/ol/fixtures/account/${PERSONA}.account.json
 
 fixtures-prod: 
-	cd ../onboard && cargo r -p onboard -- val --genesis-ceremony --autopay-file ${SOURCE}/ol/fixtures/autopay/${PERSONA}.autopay_batch.json
+	cd ../onboard && NODE_ENV=prod TEST=y cargo r -p onboard -- val --genesis-ceremony --autopay-file ${SOURCE}/ol/fixtures/autopay/${PERSONA}.autopay_batch.json
 	cp ${DATA_PATH}/blocks/block_0.json ${SOURCE}/ol/fixtures/blocks/prod/${PERSONA}/
 	cp ${DATA_PATH}/blocks/block_0.json ${SOURCE}/ol/fixtures/blocks/stage/${PERSONA}/
