+++ ol/fixtures/src/lib.rs
@@ -1,6 +1,6 @@
 use std::{fs, path::{Path, PathBuf}};
 
-use ol_types::config::{AppCfg, parse_toml};
+use ol_types::{block::Block, config::{AppCfg, parse_toml}};
 
 pub fn get_persona_mnem(persona: &str) -> String {
   let path= env!("CARGO_MANIFEST_DIR");
@@ -46,3 +46,10 @@ pub fn get_persona_toml_configs(persona: &str) -> AppCfg {
   parse_toml(buf.to_str().unwrap().to_owned()).expect("could not parse app config from file")
 }
 
+pub fn get_persona_block_zero(persona: &str, env: &str) -> Block {
+  let path= env!("CARGO_MANIFEST_DIR");
+  let buf = Path::new(path).join(format!("blocks/{}/{}/block_0.json", env, persona));
+  let s = fs::read_to_string(&buf).expect("could not find block file");
+  serde_json::from_str(&s).expect(&format!("could not parse block from file: {:?}", &buf))
+
+}
